Renamed Tiles component file to BobaTilesContainer
