+++ src/app/page.tsx
@@ -1,5 +1,5 @@
 import Navbar from '../components/Navbar';
-import BobaTilesContainer from '../components/Tiles';
+import BobaTilesContainer from '../components/BobaTilesContainer';
 
 export default function Home() {
   return (
